fix: 운영성질의 API는 info로 통일
